+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # [next]
 - Improvements performance in `LightingInterface`.
 - Improvements to check visible collisions.
+- Resolve issue [#354](https://github.com/RafaelBarbosatec/bonfire/issues/354)
 
 # [2.11.8]
 - Fix bug in `moveOnlyMapArea`