+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # [next]
 - Improvements performance in `LightingInterface`.
 - Improvements to check visible collisions.
-- Improvements in `RenderTransformer`
+- Improvements in `RenderTransformer`.
+- Update `ListenerGameWidget`.
 - Resolve issue [#354](https://github.com/RafaelBarbosatec/bonfire/issues/354)
 
 # [2.11.8]